+++ assets/set-locale-cookies.js
@@ -1,0 +1,2 @@
+document.cookie = "tz=" + Intl.DateTimeFormat().resolvedOptions().timeZone + ";path=/;samesite=lax";
+document.cookie = "locale=" + (navigator.language || "en") + ";path=/;samesite=lax";